functions/requirements.txt
azure-functions==1.12.0
azure-cosmos==4.3.0
azure-storage-blob==12.13.0